+++ tests/testthat/test-libname.R
@@ -44,6 +44,18 @@ test_that("libname() function works as expected with rds", {
   expect_equal(ncol(dat[[2]]), 9)
 })
 
+test_that("libname() function works as expected with rdata", {
+  
+  
+  libname(dat, base_path, engine = "Rdata")
+  
+  expect_equal(class(dat)[[1]], "lib")
+  expect_equal(length(dat), 2)
+  expect_equal(nrow(dat[[1]]), 10)
+  expect_equal(ncol(dat[[1]]), 9)
+  expect_equal(nrow(dat[[2]]), 2)
+  expect_equal(ncol(dat[[2]]), 9)
+})
 
 
 test_that("libname() function works as expected with xls", {
@@ -413,5 +425,53 @@ test_that("library parameter checks work as expected.", {
   expect_error(lib_info(f))
 
 
+})
+
+test_that("lib_export() function works as expected.", {
+  
+  libname(dat, base_path, engine = "csv")
+  
+  pth <- paste0(base_path, "2")
+  
+  lib_export(dat, dat2, pth, "rdata")
+  
+  expect_equal(length(dat2), 2)
+  
+  # Clean up
+  lib_delete(dat2)
+  
+  
+})
+
+test_that("lib_export() function works as expected with filter.", {
+  
+  libname(dat, base_path, engine = "csv")
+  
+  pth <- paste0(base_path, "2")
+  
+  lib_export(dat, dat2, pth, "rdata", filter = "*b")
+  
+  expect_equal(length(dat2), 1)
+  
+  # Clean up
+  lib_delete(dat2)
+  
+  
+})
+
+test_that("lib_export() function works as expected with standard_eval.", {
+  
+  libname(dat, base_path, engine = "csv")
+  
+  pth <- paste0(base_path, "2")
+  
+  lib_export(dat, "dat3", pth, "rdata", standard_eval = TRUE)
+  
+  expect_equal(length(dat3), 2)
+  
+  # Clean up
+  lib_delete(dat3)
+  
+  
 })