+++ NEWS.md
@@ -1,3 +1,12 @@
+# libr 1.2.3
+* Added "Rdata" option to `libname()` function engines.
+* Added "label" parameter to `datastep()` function.
+* Added "format" parameter to `datastep()` function.
+* Added `lib_export()` function to export a library to a different 
+directory and file format.
+* Fixed bug on `dictionary()` that was causing a warning when there 
+were no rows on the input data frame.
+
 # libr 1.2.2
 
 * Added FAQ and Complete Examples